implemented snap to grid when drag object by mouse

+++ dde-start-panel/src/ua/com/integer/dde/startpanel/image/TexturePacker.java
@@ -74,9 +74,12 @@ public class TexturePacker {
 
 	private void packCustomDir(String inputPath, String outputPath,
 			String packName) {
+		System.out.println(outputPath);
 		File[] toDelete = new File(outputPath).listFiles(new StartsWithFilenameFilter(packName));
-		for(File toDelFile: toDelete) {
-			toDelFile.delete();
+		if (toDelete != null) {
+			for(File toDelFile: toDelete) {
+				toDelFile.delete();
+			}
 		}
 		
 		Settings sets = new Settings();
